Develop

+++ app/controllers/employee/favorites_controller.rb
@@ -1,18 +1,15 @@
 class Employee::FavoritesController < ApplicationController
-
+  before_action :authenticate_employee!
+  
   def create
     @post = Post.find(params[:post_id])
-    favorite = @post.favorites.new(employee_id: current_employee.id)
-    favorite.save
-    flash[:notice] = "いいねしました"
-    redirect_to request.referer
+    current_employee.favorites.find_or_create_by(post_id: @post.id)
   end
 
   def destroy
     @post = Post.find(params[:post_id])
     favorite = current_employee.favorites.find_by(post_id: @post.id)
-    favorite.destroy
-    redirect_to request.referer
+    favorite.destroy if favorite
   end
 
 end
